Implement Debug in more places
Many of the types in fuse3 already implement Debug, but there are
exceptions.  ReplyDirectory and ReplyDirectoryPlus do not, possibly
because implementing unconditionally would force DirEntryPlusStream
and DirEntryStream implementations to also implement Debug.

This change conditionally implements Debug for ReplyDirectory and
ReplyDirectoryPlus, in the cases where the stream types also already
implement Debug.